ai-service/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning & AI
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3
joblib==1.3.2
lightgbm==4.1.0
xgboost==2.0.1

# OCR & Document Processing
pytesseract==0.3.10
pdf2image==1.16.3
Pillow==10.0.1
opencv-python==4.8.1.78
PyPDF2==3.0.1

# Data Processing
scipy==1.11.4
matplotlib==3.7.2
seaborn==0.12.2

# HTTP & API
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# Configuration & Environment
python-decouple==3.8
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Security & Validation
cryptography==41.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Time & Date
python-dateutil==2.8.2
pytz==2023.3

# File Processing
openpyxl==3.1.2
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Production
gunicorn==21.2.0
